Fix unexpected failure when messages are truncated
The library fails with [RangeError: index out of range] when a truncated
buffer is given. This patch adds checks to ensure the message is valid
and to throw a well defined err.badLength when this is not the case.

Also add tests to validate the behavior is case of bad messages.