Inherit class PDU from stream.Readable
`stream.Readable` is the standard Node.js way to deal with streams. By
inheriting `PDU` from a readable stream, we allow users to treat it as
such, therefore enabling easy writing to file, to sockets or piping.

Also add and update example usages in README.md to reflect and document
this change.