Implement streamToPDU to decode messages asynchronously
Clients (e.g. the Data project) sometimes need to pipe a stream (for instance a
socket) to generate a PDU object, but asynchronously (i.e. without waiting for
all data to be available).

This patch provides a new streamToPDU() method to decode PDU this way:

    kinetic.streamToPDU(socket, function onPDUReady(err, pdu) {
        if (err)
            // handle error
        else
            // handle pdu
    });

Moreover this function has the advantage of reading only the header and protobuf
from the stream, which means the chunk payload is left into it so another reader
can fetch and process it afterwards.