Fix class PropError badly detected by mocha
Some library (such as mocha) do not check inheritance and expect Error objects
to be really `instanceof Error`, which fails with PropError objects. This
results in mocha not displaying error message, hence making debugging harder.

Let's use a function instead of a class, so our error object are `instanceof
Error`.