PDU: Do not store chunk inside the PDU class
The chunk part is much larger than the header (9 bytes) and the protobuf message
(~ 20 bytes). To achieve performance, this chunk payload is not even supposed to
be copied ot userland.

This patch refactors the Kinetic library to avoid embedding the chunk value
inside PDU objects. Hence, clients are responsible for reading/writing chunks
from/to the source.

This patch also updates the documentation accordingly.