Add the tag needed for the put.
The tag is the integrity value for the data.
This value should be computed by the client application by
applying the hash algorithm specified below to the value(and only to
the value). The algorithm used should be specified in the
algorithm field. The Kinetic Device will not do any processing on
this value.

Update unit tests